+++ tests/integration/rpc/backend/test_evm_query_client.go
@@ -320,3 +320,11 @@ func RegisterGlobalMinGasPrice(queryClient *mocks.EVMQueryClient, height int64) 
 	queryClient.On("GlobalMinGasPrice", rpc.ContextWithHeight(height), &evmtypes.QueryGlobalMinGasPriceRequest{}).
 		Return(&evmtypes.QueryGlobalMinGasPriceResponse{MinGasPrice: math.OneInt()}, nil)
 }
+
+// Precompile
+func RegisterPrecompile(queryClient *mocks.EVMQueryClient, address interface{}) {
+	queryClient.On("Precompile", mock.Anything, mock.MatchedBy(func(req *evmtypes.QueryPrecompileRequest) bool {
+		return true
+	})).
+		Return(&evmtypes.QueryPrecompileResponse{}, nil)
+}